+++ src/Rule/Performance/DisabledCallsInLoopsRule.php
@@ -6,7 +6,9 @@ namespace Efabrica\PHPStanRules\Rule\Performance;
 
 use Efabrica\PHPStanRules\Resolver\NameResolver;
 use PhpParser\Node;
+use PhpParser\Node\Expr\Assign;
 use PhpParser\Node\Expr\FuncCall;
+use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Do_;
 use PhpParser\Node\Stmt\For_;
 use PhpParser\Node\Stmt\Foreach_;
@@ -52,8 +54,13 @@ final class DisabledCallsInLoopsRule implements Rule
             return [];
         }
 
+        $reassignedVariable = $this->findReassignedVariable($node);
+        if ($reassignedVariable === null) {
+            return [];
+        }
+
         return [
-            RuleErrorBuilder::message('Performance: Do not use "' . $functionName . '" in loop.')->tip('See https://www.exakat.io/en/speeding-up-array_merge/')->build(),
+            RuleErrorBuilder::message("Performance: Do not use \"$functionName\" in loop  to reassign variable \"$reassignedVariable\".")->tip('See https://www.exakat.io/en/speeding-up-array_merge/')->build(),
         ];
     }
 
@@ -81,5 +88,32 @@ final class DisabledCallsInLoopsRule implements Rule
         }
 
         return $this->isInLoop($parentNode);
+    }
+
+    private function findReassignedVariable(FuncCall $funcCall): ?string
+    {
+        $parentNode = $funcCall->getAttribute('parent');
+        if (!$parentNode instanceof Assign) {
+            return null;
+        }
+
+        if (!$parentNode->var instanceof Variable) {
+            return null;
+        }
+
+        $assignedVariableName = $this->nameResolver->resolve($parentNode->var);
+        if ($assignedVariableName === null) {
+            return null;
+        }
+
+        $functionParameterVariableNames = [];
+        foreach ($funcCall->getArgs() as $arg) {
+            if (!$arg->value instanceof Variable) {
+                continue;
+            }
+            $functionParameterVariableNames[] = $this->nameResolver->resolve($arg->value->name);
+        }
+        $functionParameterVariableNames = array_filter($functionParameterVariableNames);
+        return in_array($assignedVariableName, $functionParameterVariableNames, true) ? $assignedVariableName : null;
     }
 }